scripts simplification
