+++ applications/ConstitutiveModelsApplication/tests/validation/shear_traction_parameters.json
@@ -1,7 +1,6 @@
 {
     "solver_process":{
-	"solver_type"   : "constitutive_models_solver",
-	"strain_process": 0,
+	"solver_type"   : "models_solver",
 	"Parameters"    : {
 	    "properties_id": 0,
 	    "time_settings":{
@@ -10,17 +9,12 @@
 	    },
 	    "integration_settings":{
 		"integration_point": [0.75,0.75,0.0]
+	    },
+	    "strain_settings":{
+		"description": "shear strain time dependent field",
+		"deformation_gradient" : [ ["1+t*y","t*x",0.0], ["t*y","1+t*x",0.0], [0.0,0.0,1.0] ]
 	    }
-	}	    
-    },
-    "strain_process_list":[{
-	"python_module"   : "assign_strain_process",
-        "kratos_module"   : "ConstitutiveModelsApplication",
-        "help"            : "This process assigns a strain to constitutive model parameters",
-	"Parameters"      : {
-	    "description": "shear strain time dependent field",
-	    "deformation_gradient" : [ ["1+t*y","t*x",0.0], ["t*y","1+t*x",0.0], [0.0,0.0,1.0] ]
 	}
-      }
-    ]
+
+    }	        
 }
